requirements.txt
asgiref==3.3.0
Django==3.0.5
django-cors-headers==3.5.0
django-mongodb-engine @ git+https://github.com/django-nonrel/mongodb-engine@1314786975d3d17fb1a2e870dfba004158a5e431
djangorestframework==3.12.2
djangotoolbox @ git+https://github.com/django-nonrel/djangotoolbox@7e2c47e2ab5ee98d0c0f47c5ccce08e75feda48a
djongo==1.3.3
pymongo==3.11.0
pytz==2020.4
sqlparse==0.2.4
